Added Log out functionality. The new servlet needs to be called as a redirect after /adfAuthentication?logout=true
The servlet ogoutServlet will ensure again a logout from the session (which should have happend in adfAuthentication) and it will also invalidate the current token as it is no longer needed.
The servlet was configured in web.xml

OAuthEndpointServlet will set the token on the session so that it is available for the logout servlet.